Launch an EC2 worker before a node is started,...
Stop the worker when the Jenkins node is taken off-line.

The EC2 instances for the workers must be in the same region as the Jenkins master.
The master must have an IAM instance profile that permits describing/starting/stopping instances.
To specify the EC2 worker name "worker-name", label the node as "ec2:worker-name".
If no tag prefixed by "ec2:" is found, the Jenkins node
name is used.

TODO:
 - figure out how to make Jenkins pick up the extension point
 - either go back to 1-to-1 for EC2 worker and jenkins node (in scala-jenkins-infra),
   or only stop worker if all jenkins nodes that run on it are idle